This reverts the last change to get back to green.
The Extract Class refactoring says to remove the argument from #container, but I cannot
because #verse cannot depend on the underlying piece of state, it must pass _number_ -1.

Since Extract Class is always right, this means I've missed something.  If I were to say, in
English, what the first send of #container was, I might describe it as the 'initial_container'.
Hmmm.