Inject VerseVariant instead of _number_.
When _number_ is passed to initialize, THAT IS DEPENDENCY INJECTION.  Don't think
of number as a primitive, it's an object like any other object, and it was injected.

The rule is to get the right object as soon as you have enough info to do so.  Verse doesn't
need to know the name of class 'VerseVariant', it should just be injected with an object
that plays the 'variant' role.

This is beginning to look a lot like the GOF Strategy pattern.  Bottle is the context and it
uses information it has (number) to create a strategy (VerseVariant) and inject it into
Verse.  The #verse method in Bottle knows two class names (Verse and VerseVariant) but
the other classes have no such dependencies.

Bottle has-a Verse, and
Verse has-a Variant.  These relationships are composition.

All of the 'if' statements have been isolated into VerseVariant and it contains no other methods.

Now we're poised to do something really interesting....