Separate responsibility for #quantity from knowing that 0 loops back to 99.
The recipe I'm following to extract subclasses from BottleNumber aught to allow me to
remove all of the 'if' statements, but now I'm left with this ugly check of number == -1.

It makes no sense to have -1 bottles.  I don't want to make a BottleNumber-1 class. This
smells bad.

I'll fall back on SRP, and separate answering #quantity from knowing that 0 loops back to 99.
For now I'll do the dumb thing and stick the code right where I need it.